chore(ingest): reorganize unit tests (#11636)
