+++ metadata-ingestion/tests/unit/sdk/test_rest_emitter.py
